Checking in an example module, Awry, the exhibits a problem behavior. To repeat the behavior, do a checkout of Awry, then go into the directory and do a cvs update there.  You'll get a whole bunch more than just the __init__.py file on the update, despite the fact that your checkout only specified the single file, and regardless of whether the repository had changed since you did the initial checkout.
I'm investigating this behavior to see if there's a way to get just
the file.  An obvious workaround is to have only the file you want in
a directory - possibly by using a stub directory with a symlink to the
desired file, if you want to be editing the file with others in a
particular place.  However, i also want to see whether the problem
obtains with the arrangement that jim is using with the zsrc script...


git-svn-id: svn://svn.zope.org/repos/main@31233 62d5b8a3-27da-0310-9561-8e5933582275