Ok.  Experimenting with the CVS modules mechanism shows some flaws trying to knit ampersand modules into a shared directory when any of the ampersand modules designate individual files, ie,
	module -d target/loc origin/dir file ...

When that arrangement is used, updates done immediately after the
checkedout behave peculiarly - in some cases the updates bring in
files and directories excluded from the checkout (defeating the
specific-file designation), and in some cases repeated updates
perpetually show unedited files as being modified.

A workaround that seems to be working is to create directories
containing symlinks to only the desired individual files, and knitting
in those directories in their entirety.  I'm using this to create an
entire Zope module, and it seems to behave properly after a bit of
testing.


git-svn-id: svn://svn.zope.org/repos/main@31234 62d5b8a3-27da-0310-9561-8e5933582275