Fixed a small typo, and refined the module docstring link example, in order to do a checkin to exercise the CVS repository mirroring.  Might as well include my last checkin message, with some substantial stuff:
Links are now recognized whether or not the candidate strings are
terminated with punctuation before the trailing whitespace.  The old
form - trailing punctuation then whitespace - is preserved, but the
punctuation is now unnecessary.

The regular expressions are a bit more complicated, but i've factored
out the common parts and but them in variables with suggestive names,
which may make them easier to understand.


git-svn-id: svn://svn.zope.org/repos/main@16050 62d5b8a3-27da-0310-9561-8e5933582275