When reexecing, append to the existing log, instead of overwriting it, so we preserve the primary comments.
TO BE DONE - the log file is currently shared, and concurrent checkins
will stomp eachother, interspersing or overwriting their log entries.
This is not desirable - but then again, we don't want a whole bunch of
logs left around, and this reuse is an easy way to limit that.  A good
solution short of painful, elaborate things like locking or some
reaping of old logs is not apparent - so we're sticking with quick and
dirty for now.


git-svn-id: svn://svn.zope.org/repos/main@31247 62d5b8a3-27da-0310-9561-8e5933582275