Special provisions to accomodate relocation of actual webdav directory, with CVS problems due to the fact that its effective location stays the same.
People who had the old location in their bookkeeping normally would
have that rectified when they do an update, but since the webdav dir
would be showing up in the same location, just from a different
repository dir, the clearing process fails.  To get around that, we
left a symlink in the old repository dir, to satisfy the old,
unreconcilable checkout bookkeeping.

This would mean that new checkouts get the incorrect bookkeeping,
except we're using a '-l' in the entry of the old origin ("GreyThing")
to prevent recusion into it (fortunately, it has no other subdirs), so
new checkouts will get the right bookkeeping.  The addition of the 'l'
is the whole change of this checkin!-)


git-svn-id: svn://svn.zope.org/repos/main@31264 62d5b8a3-27da-0310-9561-8e5933582275