Changed INSTANCE_HOME to be = to SOFTWARE_HOME + 'var'.
Fixed a doc string that said S_H was the same as I_H

Added a Unix path to cheapen the experience and remove that nasty DOS
path.

If the log file doesn't exist, create instead of failing

Updates an INSTANCE_HOME + 'ZServer.pid' file with the process
id.  To convieniently kill ZServer.  yes this work with wondows.


git-svn-id: svn://svn.zope.org/repos/main@16296 62d5b8a3-27da-0310-9561-8e5933582275