Added WARNING as an alias for PROBLEM.
Added a format_exception method.  This is essentially the method
used by the ZPublisher exception handler and has the nice feature
of showing object __names__ and __traceback_info__.  It is useful
in back-end log handlers *and* in other applications that want to
display tracebacks.

Added output of exception info to the stupid logger.


git-svn-id: svn://svn.zope.org/repos/main@16742 62d5b8a3-27da-0310-9561-8e5933582275