ExternalMethod.__call__(): Fixed DevelopmentMode behavior so it didn't reload extenal method files every time.
The problem was that the entire os.stat() record was being compared,
and that includes the access time, which of course changes every time
the method is loaded - self fulfilling prophecy, and all.

Instead, we now just stash and compare the modification time
(ST_MTIME) field, and get rid of the extraneous (expensive) reloads.


git-svn-id: svn://svn.zope.org/repos/main@17250 62d5b8a3-27da-0310-9561-8e5933582275