The following scenario could lead to strange and serious errors:
  - Restart or pack Zope, causing an index to get written

  - Undo some transactions

  - Shutdown Zope ungracefully (e.g. kill)

When Zope comes up, the index points to undone records for some
objects. This can lead to all sorts of problems. :(

To overcome this problem, we now remove the index after an
undo.

This increases the likelyhood that it will take longer to restart
next time, but reduces the chance of a bad index.


git-svn-id: svn://svn.zope.org/repos/main@898 62d5b8a3-27da-0310-9561-8e5933582275