Some slight refinements (indicating directory where chmod command is occuring, during verbose operation) in my quest to debug my script over what turned out to be an rsync (2.1.1) bug.
Re the rsync bug.  In certain special circumstances the homedir of the
account servicing the command on the receiving host was being chmod'ed
g+w (and who knows what else), despite the fact that the directory was
explicitly mentioned nowhere in the command.  Anyway, with 2.3.1 no
running on the mirror host, the misbehavior has ceased.  (The problem
particularly showed because ssh doesn't accept the homedir of the
remote account being g+w.  Consequently it worries me a bit to think
that they may have only caught this because it broke something so
overtly - what about the chmods, or god knows what, that aren't
exposed??  Oh well.)


git-svn-id: svn://svn.zope.org/repos/main@31314 62d5b8a3-27da-0310-9561-8e5933582275