A recent change caused HTML files to have their headers and footers replaced with var tags for standard headers and footers.  While this is sometimes very useful, it is sometimes disastrous.  This feature is now enabled with the -D option.
Handling of files with a '.dtml' extension was broken. These are
now handled correctly. Files with '.dtml' suffixes now get uploaded as
methods.

A workaround for old sites that had a bug in Document (aka DTMLMethod)
upload has been disabled. This workaround is now enabled when the -9
option is used.

Added logic to try and figure out a path to ZPublisher if one was not
provided.

Fixed a bug in misshandling binary files on windows.

Added an option, -I, to automatically add an index_html method that
redirects to index.html or index.htm.


git-svn-id: svn://svn.zope.org/repos/main@3934 62d5b8a3-27da-0310-9561-8e5933582275