Added a separate call to check the cache size so that we only check when a transaction is about to be written. It would be bad if we "flipped" the cache twice during a transaction, as we would have incomplete transaction data.  Note that this means that we could exceed the target cache size if the data for a transaction exceeded the target size (/2).
Changed the way that the storage is closed. Our close method may be
called from a separate thread than the async main loop. We only want
the async main loop to close our _call (rpc) object. Instead of
closing it directly, we call closeIntensionally(), which, in return,
requests that the async main loop closes it, via the select trigger.


git-svn-id: svn://svn.zope.org/repos/main@975 62d5b8a3-27da-0310-9561-8e5933582275