Added ClientStorage option to wait for a connection to the server on start up, with the default being to wait for the server.
The intent was that it should be possible to bring a client up even
if it can't talk to the server, however, this is a little bit risky if
we need to start them both.  Eventually, we need to be the start-up
code be smart enough to wait if we can't serve data requests during
startup, but to proceed if we can.


git-svn-id: svn://svn.zope.org/repos/main@1048 62d5b8a3-27da-0310-9561-8e5933582275