Folded in a number of changes from Python CVS.
Added a new "base" type, IOobject for which most of the
operations are defined.  This will, hopefully clarify
some of the logic.

Added close test to raise proper error when operations
are performed on closed StringIOs.

Added a position argument to the truncate method.

Added a size argument to readline.

Added PyArg_Parse calls for methods that don't take arguments to
make sure they don't take arguments.

Guido-fied some of the indentation.


git-svn-id: svn://svn.zope.org/repos/main@18516 62d5b8a3-27da-0310-9561-8e5933582275