Merging changes from Zope-2_2-branch:
rfc1123_date(): Put back in the check for defaulted timestamp, so we
don't get exceptions in that case.  Fixes failure on writing ftp
documents, at least.

Looks like the 'format' parameter is unused and inappropriate, but i'm
leaving that for someone in the know to repair.  (They should pull out
my 'XXX' comment/note, at that time.)


git-svn-id: svn://svn.zope.org/repos/main@18571 62d5b8a3-27da-0310-9561-8e5933582275