Change RuntimeError to SGMLParseError, which carries line number and offset information.
Add support for SGML declaration syntax (<!....>) to some reasonable
degree.  This does not support everything allowed in SGML, but should
work with "real" HTML (internal subset in a DOCTYPE is not handled).
The content of the declaration is passed to the .handle_decl() method,
which can be overridden by subclasses.


git-svn-id: svn://svn.zope.org/repos/main@19559 62d5b8a3-27da-0310-9561-8e5933582275