Move compile() into the DummyEngine() class (deleting our DummyCompiler class), and make sure to call it everywhere it is needed (in particular for attribute replacements).
Emit a new opcode, "mode", with an argument "html" or "xml", depending
on the value of the xml argument to the constructor.  (This requires a
matching checkin of TALInterpreter.)  (This comment was mistakenly
added to the previous checkin of DummyCompiler.py.)

Don't append to self.program directly; always call self.emit().


git-svn-id: svn://svn.zope.org/repos/main@19679 62d5b8a3-27da-0310-9561-8e5933582275