Add opcode implementation for "mode".
Properly implement tal:replace="structure ..." combined with
tal:attributes="...", both in HTML and in XML mode.  This is pretty
tricky, and requires instantiating a parser with a modified code
generator which suppresses any TAL/METAL interpretation but does the
attribute replacement on the first encountered start tag as required
by the TAL 1.0 specs.


git-svn-id: svn://svn.zope.org/repos/main@19681 62d5b8a3-27da-0310-9561-8e5933582275