BaseFramework => MinimalBaseFramework
MinimalBaseFramework.setUp(): Wrap the creation and opening of the
database in a try/except so that we guarantee that tearDown() is
called if there are any errors in database creation.

tearDown(): Added this to make sure that the test directories are
removed when the test is done.

FullBaseFramework: New class to mirror the same tests for the Full
database.  Based on MinimalBaseFramework.  This could use some
factorization.

DBHomeTest => MinimalDBHomeTest

FullDBHomeTest: New class to mirror the same tests for the Full
database.  Based on MinimalDBHomeTest.  This could use some
factorization.

Added __main__ section so this test can be run standalone.


git-svn-id: svn://svn.zope.org/repos/main@19751 62d5b8a3-27da-0310-9561-8e5933582275