Moved Python 1.x compatable asyncx to their own package that takes care of installing them when necessary (not for Python 2.0)
Now import asyncore and asynchat as top-level modules, rather than
from the medusa package. This makes it easier to use our custum
version for Python 1.5 and to use the standard versions for Python
2.x.

Note that the Python 2.1b1 and 2.1b2 releases had a broken asynchat
that causes these changes to break Zope. The Python cvs and earler and
later versions are OK.


git-svn-id: svn://svn.zope.org/repos/main@19805 62d5b8a3-27da-0310-9561-8e5933582275