Fix bug reported by Richard Jones in ZPT(19).
Also fix the error for a duplicate fill-slot (which was erroneously
reported as a duplicate slot *definition*).

Question arises, should we also complain about duplicate define-slot,
or is it okay to use define-slot with the same name more than once?
This would be similar to "#define X(a) (a+a)" in C; it does point out
that define-slot isn't a great name...


git-svn-id: svn://svn.zope.org/repos/main@19807 62d5b8a3-27da-0310-9561-8e5933582275