Title:  Results class optimization
At: http://classic.zope.org:8080/Collector/Collector/2141/sview

Submitter:  John Eikenberry

Email:  jae@kavi.com

Description: We have a large query implemented via a SQL Method. We've
been messing with it to get the most speed out of it, and noticed a
large hit in the __init__() method on the Results class. We found that
with this simple tweak (see included patch) we got around a 180% speed
up. Not a ton, but pretty good for changing 3 lines of code.


git-svn-id: svn://svn.zope.org/repos/main@19832 62d5b8a3-27da-0310-9561-8e5933582275