Comment out the debugging crutches.
_zaprevision(): Fixed some silly typos.

pack(): Adopt from FileStorage, the conversion scheme of time floats
to TimeStamps.  Also use set_range() on the txnoids cursor because
it's not likely that the timestamp (i.e. revid) will actually exist,
and we just want the one just before the specified time.  Use prev()
instead of prev_dup() on the cursor.


git-svn-id: svn://svn.zope.org/repos/main@19978 62d5b8a3-27da-0310-9561-8e5933582275