Adding request log profiler.
Usage: requestprofiler.py filename [--sort=spec] [--top=n] [--verbose] [--help]

Provides a profile of the detailed (-M) Zope request log.

Each line in the profile indicates information about a Zope method (URL)
collected via the detailed request log (the -M log).

'filename' is the path to the '-M' log that contains detailed request data.

If a 'sort' value is specified, sort the profile info by the spec. The sort
spec may be any of 'hits', 'hangs', 'max', 'min', 'mean', 'median', or 'total'.
The default is 'total'.  The sort order is decending unless indicated.

  'hits'       -- the number of hits against the method
  'hangs'      -- the number of unfinished requests to the method
  'max'        -- the maximum time in secs taken by a request to this method
  'min'        -- the minimum time in secs taken by a request to this method
  'mean'       -- the mean time in secs taken by a request to this method
  'median'     -- the median time in secs taken by a request to this method
  'total'      -- the total time in secs across all requests to this method
  'url'        -- the URL/method name (ascending)

If the 'top' argument is specified, only report on the top 'n' requests in
the log (as per the sort).  The default is 10.

If the 'verbose' argument is specified, do not trim url to fit into 80 cols."""


git-svn-id: svn://svn.zope.org/repos/main@19986 62d5b8a3-27da-0310-9561-8e5933582275