Remove many unused default argument definitions used as speed optimizations.
Many methods had default arguments that were used to bind a global or
builtin name, like type, None, or type(''), as a local name.  In many
cases, the names bound were not referenced in the method body
(presumably they used to be referenced).

commit_sub(): In this one case, delete the default arg anyway, because
    it's only used once.


git-svn-id: svn://svn.zope.org/repos/main@1331 62d5b8a3-27da-0310-9561-8e5933582275