With these changes i can now easily convert existing WikiForNow wikis to CMFWikis - i'll check in the script to do that momentarily.
.__init__(), makeCMFWikiPage(): Cleaned up argument names, and
rectified so we actually get the initial content passed to
DTMLDocument.__init__() properly.

addCMFWikiFolder(): Attribute FrontPage as SandBox page parent (a
bit of a kludge, but it's ok to have one special case out of two
cases, ay?)

Added some stray attributes to the official properties:

 ._subowner (plus, itemizing SUBOWNER_TYPES so we can have a select
            list on the properties page)
 .comment_number

This makes it easy to automate copying all the properties when
migrating old WikiForNow ZWiki Pages.


git-svn-id: svn://svn.zope.org/repos/main@33630 62d5b8a3-27da-0310-9561-8e5933582275