Implemented a concurrent-edit safety belt for Document objects, like the one jim added to ZWiki, to prevent edits that accidentlly stomp concurrently occurring edits.  This implementation has a refinement so a user can save a document repeatedly using the same seat belt (eg, without revisiting the document after every save).  This allows for common webdav or emacs-based ftp editing (or use of the web browser back button to reedit) usage patterns, while still providing conflict prevention.
It's possible we'll eventually be handling metadata headers in FTP and
webdav differently, which may mean taking a different tack on this
functionality (or discarding it).

._last_safety_belt_editor, ._last_safety_belt, ._safety_belt: Class
instance vars to track the safety belt state.

._edit(): Added safety_belt parameter, for web case where the safety
belt is not passed in the headers.  Extracting the safety belt info
from the headers or param, and using ._safety_belt_update() to
validate - if it fails, then raise an 'EditingConflict' exception with
an informative message.

.getMetadataHeaders(): Provide our own version of this method,
appending the safety belt header to the generic
(DefaultDublinCoreImpl) one.

.SafetyBelt(): Expose the safety belt value for consumption from the
web edit form.

._safety_belt_update(): Enforce safety belt rules, returning 1 when
satisfied, 0 when not.

.PUT(): Catch 'EditingConflict' exception that would be caused by
safety belt conflict trigger, aborting the transaction and returning a
450 FTP failure.


git-svn-id: svn://svn.zope.org/repos/main@33639 62d5b8a3-27da-0310-9561-8e5933582275